Moved and renamed simplification methods to simplification file.
